Move fine-grained NodeKind to ReferenceDependencyKeys.h
